tribalance-proxy/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tribalance-proxy"
version = "0.1.0"
description = "Lambda Function URL proxy between TriBalance frontend and AgentCore Runtime"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
    "boto3>=1.42.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "ruff>=0.7.0",
    "httpx>=0.28.0",
]

[tool.hatch.build.targets.wheel]
packages = ["infra"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"
